TASK: Split up Projection into Projector and Finder
